http/requirements.txt
Flask==0.12.2
flask-restplus==0.9.2
google-auth
google-auth-httplib2
requests
